requirements.txt
pip==21.0
setuptools==57.5.0
wheel==0.37.0
pandas
pandas_ta
python-binance
numpy==1.23.5
apscheduler
requests
matplotlib
ipykernel
gym==0.20.0
ray[rllib]==1.13.0
torch
pyyaml
ta
aiohttp
raylib-py
flask
gputil
pytest